Add script to build compatible disk image
The resulting disk image can be used in Peter De Wachter's Oberon Emulator
as well as the Java Oberon Emulator (it will not work in the JavaScript
version and will crash occasionally in the Java version without floating
point support).

When having built the image following OberonFromScratch.Tool, use a Hex
Editor to strip the first 1KB of the image away (as the ImageTool will put
the directory root in sector 1, but it has to be in sector 0 so that the
image is detected as a filesystem-only image).