Add 32-bit address space support
This is optional, and enabled by some unused configuration words at the
end of the boot loader image. If enabled, the address space width is 32
bit and the first (configurable) X MB of it are mapped with real RAM. The
position of the display framebuffer and the ROM are also configurable.
Memory mapped IO remains to occupy the last 64 bytes.

BootLoadBigMemoryConfig can be used to patch a bootloader (both the
configuration words and the actual code itself) compiled from BootLoad.Mod
so that it will use X MB of RAM, Y MB of it for modules space and stack
(by default half of X) and map the framebuffer at 0FCE00000h (50MB from
the end of the address space) and the ROM at 0FFF00000h (1MB from the end
of the address space).

make_bigmem_image.sh can be used to prepare Oberon sources for building a
disk image that uses all the special emulator features (including color
display and clipboard integration in the viewer menus) and that works with
a boot loader patched by BootLoadBigMemoryConfig.