Add support for limiting emulator features
Features can be limited to predefined feature sets (like NATIVE or C or
JS) or to a more fine-grained list of features that should be allowed.

Useful for testing compatibility of generated disk images with other
emulators/environments.

In case a disabled feature is used, the CPU stops executing and an
exception message is printed on the console.