Another approach at larger screen sizes
As implemented by @pdewacht's Oberon Emulator, this approach leverages the
fact that the memory layout is adjustable and therefore the image memory
can be larger than 96KB, and just increases the line span between lines.
To make it possible to detect this kind of emulators, it uses a different
magic value for communicating the screen size.

To use this approach, give a negative value as the screen width and (in
case you use LimitFeatures) make sure that the NEW_DYNSIZE_GRAPHICS
feature is enabled. It is your own responsibility to use a boot loader
image that configures the memory layout so that there is enough space for
your framebuffer.

Also update the Display.Mod used for the compatible disk image to the
version that supports this approach.