Ignore spaces in Batch.VerifyLog
When you print numbers to the Log, they will be padded by spaces - so even
if you ignore/fold digits, if the number changes enough, it will still
change the number of spaces used for padding. Therefore ignore spaces too
(and adjust the OberonFromScratch.Tool accordingly) to make the build even
more reliable.