Run Oberon Experimental in the JavaScript emulator
Oberon Experimental are "Various experiments with the functionality, the
module structure and the boot process of Oberon 2013" by
@andreaspirklbauer (see
https://github.com/andreaspirklbauer/Oberon-experimental for more
details).

Use the Java emulator while building, and apply the attached patch to the
files when you build stage 2. The resulting Oberon Experimental will run
in the JavaScript emulator (as well as in the Java Emulator).