Add another backbuffer to screen canvas.
Writing to the memory-mapped graphics memory changes pixels only in this
back buffer, and an asynchronous event will update the dirty area from
this back buffer into the screen canvas. This needs an additional (maximum
of) 3 MB RAM (4-byte RGBA for 1024x768 pixels), but behaves a lot faster
for large screen updates (which happen a lot in Experimental Oberon).