Refactor risc into RISCMachine
This eliminates the system as a global singleton.  If you wanted to
simulate multiple sytems, for example, you can now instantiate a RISC
machine for each node.  (NB: Each system will still need a driver, and
currently WebDriver is only meant to provide a frontend for a single
emulator instance.  However, a command-line driver is is forthcoming.)