Start work on disk sync
I got sick of having to access the emulator from localhost instead of
running it through a file:// URI.  (Browsers throw a privilege violation
exception when trying to use the associated canvas methods on images
loaded from file:// URIs.)

The emulator's new disk loading routines are going to allow loading from
a raw image, rather than requiring that it be "cooked" into a PNG.
Additionally, there will be a file selector so you'll be able to select
your own file to use, and there will be a way to export the disk image.

I might integrate support for vfsync.org, too.  In that case, the
selector in index.html will go away, and the emulator will launch from
publicly hosted images.