Refactor register access
The idea here is that all register operations will go through
`setRegister` and `getRegister`.  This is to make debugging the JS
implementation easier, since you'll be able to set a breakpoint to
observe all (or some) register operations.

There is a minor performance penalty, but our benchmark is 25MHz.

In the future, this will also be useful for implementating
break-on-register-change in the (yet-to-be-written) Oberon/RISC5
debugger.