Add doctype, etc to avoid warnings
This also fixes a bug where there were two extra close and open tags for
the head and body elements, respectively.

Note that the doctype triggers standards mode, which means there is work
ahead to retweak the layout.  This should have been done sooner for the
emulator, to avoid the headaches when trying to style the revamped UI...