Fix inline style assignments for standards mode
When we introduced the doctype, the page was being rendered in Firefox
and Chrome in quirks mode.  The doctype enabled standards mode, which
means some things break that would have worked in quirks mode.  Style
values with no unit are among the things that will break.  (Whereas in
quirks, it defaulted to px).  We now specify px if that's what we mean.