Add "memory alligator" patch
As written in
<http://lists.inf.ethz.ch/pipermail/oberon/2017/011016.html>, Oberon's
memory allocator has a bug that if an allocation of a large block fails,
it will nevertheless try to partition the resulting null pointer into
smaller parts if a small block has been requested, causing memory
corruption that ultimately leads to a crash.

Fix this by returning the null pointer in the smaller allocator
functions if the big one returns a null pointer.