Refactor link status/queuing (for readability)
Among my notes about the in-browser emulator are two comments:

  1. The getStatus implementation needs to be cleaned up for readability
  2. There needs to be support for queued demand transfers

Michael also emailed me last week about his change to fix the latter.  I
want to take care of the former first, however, because it affects the
second.  So these changes are meant to take care of (1).  The changes
for (2) will build on top of that.