Fix ever-increasing popup size
Due to the +1 hack in togglePopup, every time a menu is opened, it will
increase in width by 1 px in Gecko/Blink/WebKit, something most
noticeable in the Transfer menu.

The +1 hack was originally in place due to layout issues in IE10, which
without the hack will place the Copy button on the line below the text
input.  The +1 hack fixes this for IE, and still leaves things looking
acceptable in other browsers.  (In reality, it leaves a small gap between
the edge of the Copy button and the parent menu's frame/box object in
other browsers, but this was determined to be a small tradeoff to get IE
working.)

With this change, the width will be computed only once--the first time
the menu is opened.