Refactor image file readers
These changes are important for being able to eventually ditch the
current index.html in favor of using the controlbar UI to select a
system and options.

The readers have been refactored to conform to a common interface that
allows for lazier reading--initializing the reader without immediately
beginning a read + boot.