Add ability to load default ROM
These changes are yet still work that will allow us to kill the current
index.html in favor of just linking directly to emu.html and booting up
by choosing an image from the system selector.  (NB: As of this change,
this is not yet possible, but it is possible use the Load menuitem and
select a simple disk--i.e., one where the disk contents start at 0 and
which doesn't contain a phantom ROM sector embedded in the payload.
It's still possible to run one of the premade, PNG-encoded images by
specifying it in URL parameters as before, of course.)

The default ROM provided here is derived from JSBootLoad.Bin.