Add config.json to list available images
There are a few benefits to this:

1. This allows site-specific configuration that doesn't require editing
   the emu.html source by hand to add available images
2. It removes control from third-parties to make arbitrary GET requests

The mitigations in #2 are actually against a low-risk attack vector,
(assuming a sane website and a properly configured webserver--i.e., one
for which GET requests are idempotent and free of side-effects, as they
should be).  It is, however, a convenient benefit to the main motivation
for these changes, which is #1.

What #1 means is that anyone can "install" an instance of the JS
emulator on their server, and they need only place the available images
alongside the emulator's own files and make sure they're listed in
config.json.  The alternative would be edit emu.html by hand to list the
available images.  That's hazardous both because the HTML is somewhat
involved (and therefore easy to bungle) and because emu.html is a
source-controlled file (making dealing with local edits a bit onerous).

This is similar to the way jslinux works <https://bellard.org/jslinux/>.