Add emulator size options to settings menu
This change effectively makes index.html obsolete; anyone can navigate
directly to the emulator, even without special URL parameters, and
select the preferred size and system image from there.  The upshot is
that it's almost trivial now to embed on any webpage a widget that acts
as an Oberon sandbox to play around with.