Add support for larger RAM to JS emulator
This is done by having a magic string "MB=x" at the end of the boot
loader. Unlike the previous large RAM support in the Java emulator,
memory mapped IO and display memory are placed inside this RAM (not as
virtual RAM after the physical RAM). The same mode is also added to the
Java emulator to make testing easier.