Add Autosave to JS emulator
When enabled, current disk image content is stored (sector by sector)
into local storage, whenever it is written to.

When loading the autosaved image, autosave is enabled again as a
convenience function. When disabling autosave, you get the option to
delete the autosave image.

Multiple autosave images are not supported. If you need to use a
different image, save the previous autosave image and delete it.