Add HostFS support to Java emulator
When using HostFS, you need to replace FileDir and Files by special
HostFS versions. As the filesystem semantics of Java are different than
the ones of Oberon (e.g. on Oberon you can unlink a file name and then
link it to a new name afterwards), unregistered files are represented by
temp files starting with a ~, which are deleted on emulator startup.