Add scripts to rename all modules of the system
Used to create a barebones system where all file names start with R5.

This can be useful when trying to port/bootstrap another system without
running the risk that a just recompiled file clobbers a file of the
system you are working on, while still making it possible to load and
use those compiled modules if desired.