Add low-level trap viewer
When a trap or abort occurs, a box is drawn to the center of the screen,
containing trap number, module name, position and memory address.

When L+R mouse button have been pressed and released, forwards the trap
to the previously installed trap/abort handler.

That way, when a repeated trap occurs inside low-level modules like
Texts, the trap information will still be displayed and can be
investigated manually after a hard reset.