Fix delay when closing Java emulator
When the virtual CPU is currently waiting for the next task, closing
could have been delayed as long as the start time for that task is
reached. As the bootloader sets an "infinite" delay while waiting for
RS232 input, emulator does not close at all when trying to close it
while booing via serial link.