Add defragmenter module
Useful to shrink/trim emulator images from within the emulator.

Native trimming was not implemented since DISCARD (CMD38) is only
supported for SDHC/SDXC (so not for small SD cards) and also not
accessible via SPI. Instead, write a magic sector to the card, which can
later be used to trim the image (trim_defragmented_image.sh).