Add paravirtual real-time clock support
At startup, the emulator writes the current tick count and the
real-time-clock (in local time) to addresses starting at 0x10000.

A patch to Kernel.Mod picks up the values and initializes the real-time
clock, also taking care of dynamically updating the real-time clock when
it is accessed (to avoid a background task that wakes up the virtual CPU
every second).