Dual serial port support for Java/JS emulators
That way, you can have both a serial port between two VMs, as well as
PCLink functional in both VMs, at the same time.

The serial status register will use two more bits for the second serial
port status, and writing to the serial status register is used to select
a port for reading/writing the serial data register.

Also add an option for the JS emulator to suggest a different disk image
for the other end of the serial link.

A patched PCLink module is also included which will automatically use
the second serial port, if present.