JS: Dynamic RAM size support
For supported images, when resetting the emulator, the RAM size (and
color support) can be chosen from 1MB to 64 MB.

This information is stored as magic value just before the Palette start
and can be used by the bootloader from there.

(Note that existing Color images need their boot loader patched so that
it indicates color support. Apart from that, this change should be
backwards compatible).