Add small webserver for JavaScript WizNet support
A small Java webserver (based on Jetty) is used to provide a websocket
to TCP/UDP socket "proxy". That way, WizNet can be emulated in the
JavaScript emulator.

This only works for the "offline version" (which is served by the
webserver as static resource as well), not for the "online version".

The webserver requires Java 8 or higher and Jetty 9.4.33.