Add support for JavaScript Clipboard API
It is not very widespread yet (fully supported in latest Chrome, writing
clipboard supported in latest Firefox), but when it is supported it is
very convenient.

Due to the low browser support, and as it asks for permission, use it
only if a Clipboard command has been called before the "virtual"
clipboard has been opened. Otherwise use the virtual clipboard as
before.