Fix automatic build of all JS images
Streamline the image generation scripts by providing a "v2" option to
some helper scripts. Add an option to CommandLine runner to sleep
between each command (seems there is some race condition with the
Garbage Collector otherwise). I know that this is not a nice solution,
but it works suprisingly well.