Java: Add hardware enumerator stub
This is a temporary solution to fix a "cyclic dependency" between
uncommitted changes on my disk.

I am working on a larger rework of the Java emulator with more intuitive
command line parsing, hardware enumerator support and more. However,
this is not ready for publication yet, and I also plan to update the JS
emulator first and then make sure the Java emulator is compatible with
the JS emulator.

Before updating the JS emulator, I need to update all the JS images as
some of the old quirks of the JS emulator (used by the old images) will
go away. Yet, the build system for the JS images currently uses the Java
emulator.

Therefore, this small commit adds some duct tape to the Java emulator to
- support the new command line syntax for just the case used by image
  building
- support just enough of the hardware enumerator so that images can be
  successfully built