Fix deadlock while indexing files
Editing fils while indexing a large project can lead to a deadlock: When
a module is finished being indexed, its dependent modules are checked to
find any unused dependencies. When initial indexing of that file has
already finished, but it is queued again for re-indexing due to recent
changes in the file. In that case, accessing the file will block until
re-indexing has finished, which will never happen as the re-indexing has
to wait for the dependency checking to finish.

Work around this issue by changing how error markers and dependency
usage are determined: When these actions are called from the indexer
thread, stop them whenever a file needs to be accessed that is still
pending indexing; in that case, do not report the error markers to the
editor. When the re-indexing has finished, the error markers and
dependency usage will be analyzed again anyway.