Raw JavaScript Event keyboard emulation
This feature of the JS emulator provides raw access to the underlying
KeyboardEvents to the emulator, so that it can support more
sophisticated keyboard drivers than the paravirtual keyboard can.

Document the hardware enumerator extension, implement in the JS
emulator, and provide a patch for the Oberon system to use.