Java: Support different PNG encoding layouts
Until now, the PNG always had a width of 1024 pixels and the height
corresponded to the number of 1K sectors. Since Firefox does not want to
decode PNG files into `<canvas>` that are higher than 32768 pixels, make
the width dynamic (any multiple of 1024 pixels) and put multiple sectors
into one row.

Also, avoid an exception when trying to encode a PNG with the
`--no-padding` option.